+++ packages/core/src/db/index.ts
@@ -2,6 +2,7 @@ import type {
   PublisherConfiguration,
   Seat,
   Subscription,
+  SeatCreationContext,
 } from "../common/index";
 
 export interface Repository {
@@ -22,6 +23,11 @@ export interface Repository {
     byEmail?: string
   ): Promise<Seat[]>;
   replaceSeat(update: Seat): Promise<Seat>;
+  createSeat(
+    seat: Seat,
+    subscription: Subscription
+  ): Promise<SeatCreationContext>;
+  deleteSeat(seatId: string, subscriptionId: string): Promise<void>;
   getSubscription(id: string): Promise<Subscription | undefined>;
   getSubscriptions(publisherId: string): Promise<Subscription[]>;
   createSubscription(